# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AninfPush API Documentation",
        "description": "API microservice for managing multi-channel messaging (Email, SMS, WhatsApp) with Keycloak authentication",
        "contact": {
            "email": "support@aninfpush.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://aninfpush.com/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/dashboard/stats": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get dashboard statistics",
                "description": "Returns overall statistics for messages, businesses, and costs",
                "operationId": "d87dd1b4d8881af3798cb3a56f1740ee",
                "parameters": [
                    {
                        "name": "business_id",
                        "in": "query",
                        "description": "Filter by business ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date for filtering (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date for filtering (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Operation successful"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_messages": {
                                                    "type": "integer"
                                                },
                                                "sent_messages": {
                                                    "type": "integer"
                                                },
                                                "delivered_messages": {
                                                    "type": "integer"
                                                },
                                                "failed_messages": {
                                                    "type": "integer"
                                                },
                                                "pending_messages": {
                                                    "type": "integer"
                                                },
                                                "total_cost": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "messages_by_type": {
                                                    "type": "object"
                                                },
                                                "messages_by_status": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/dashboard/recent-messages": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get recent messages",
                "description": "Returns a list of the most recent messages",
                "operationId": "0a7f6db93c54d634fd38f72e91549932",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of messages to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "business_id",
                        "in": "query",
                        "description": "Filter by business ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Operation successful"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Message"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/dashboard/message-trends": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get message trends",
                "description": "Returns message trends over time grouped by type",
                "operationId": "9426d5b564bd1a401282287ee5c03c94",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Time period for trends",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "week",
                            "enum": [
                                "week",
                                "month",
                                "year"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Operation successful"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "date": {
                                                        "type": "string"
                                                    },
                                                    "message_type": {
                                                        "type": "string"
                                                    },
                                                    "count": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/dashboard/cost-analysis": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get cost analysis",
                "description": "Returns cost breakdown by type and date",
                "operationId": "e573478a0e6faa22dccf33e409a93c5e",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date for filtering (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date for filtering (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Operation successful"
                                        },
                                        "data": {
                                            "properties": {
                                                "by_type": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "by_date": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "total_cost": {
                                                    "type": "number",
                                                    "format": "float"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Business": {
                "title": "Business",
                "description": "Business model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Acme Corporation"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "contact@acme.com"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "+237670000000"
                    },
                    "country_code": {
                        "type": "string",
                        "example": "+237"
                    },
                    "website": {
                        "type": "string",
                        "example": "https://acme.com"
                    },
                    "description": {
                        "type": "string",
                        "example": "A leading technology company"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main St"
                    },
                    "city": {
                        "type": "string",
                        "example": "Douala"
                    },
                    "country": {
                        "type": "string",
                        "example": "Cameroon"
                    },
                    "timezone": {
                        "type": "string",
                        "example": "Africa/Douala"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "suspended"
                        ],
                        "example": "active"
                    },
                    "verification_status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "verified",
                            "rejected"
                        ],
                        "example": "verified"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
            "Message": {
                "title": "Message",
                "description": "Message model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "business_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "message_type": {
                        "type": "string",
                        "enum": [
                            "email",
                            "sms",
                            "whatsapp"
                        ],
                        "example": "email"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "queued",
                            "sending",
                            "sent",
                            "delivered",
                            "read",
                            "failed",
                            "cancelled"
                        ],
                        "example": "delivered"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T00:00:00Z"
                    },
                    "business": {
                        "$ref": "#/components/schemas/Business"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter Keycloak JWT token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}